Move and break up the scheduler
