+++ src/dispatch/error.py
@@ -1,0 +1,3 @@
+class IncompatibleStateError(Exception):
+    """Coroutine state is incompatible with the current interpreter
+    and application revision."""
